Rework `tag_variables()` to `label_variables()`
This commit starts the migration from using the
data.frame level attributes for tags (that is
`attr(x, "tags")`) to using the `label` attribute
at the variable level (that is `attr(x$var,
"label")`). This is the start of replacing the
entire backend for the package as is currently
proposed in #32.