Refactor `restore_tags` with downstream effects
This commit untangles multiple functions, joined
primarily by their relation to the refactor of
`restore_tags` to `restore_labels`. The refactor
includes the entire removal of `prune_tags` (which
is made redundant by the column level labelling).
The lost_tags_action is moved to the
`restore_labels` function. The `square_bracket.R`
file is updated to ensure that the operations work
as expected. Unit tests are included for the new
code, where necessary.