Refactor `tags_types` + `validate_[types/labels/datatagr]`
This commit removes `tags_types` and incorporates
it directly into `validate_types`. `validate_types`
works on specified variables now. `validate_labels`
no longer requires everything to be labelled, only
for those labelled to be valid. There is now the
option to specify a type for variable without a
label, allowing for more flexibility in usage.